+++ jobbergate-core/tests/auth/test_handler.py
@@ -1,0 +1,395 @@
+"""
+Test the utilities for handling auth in Jobbergate.
+"""
+from unittest import mock
+
+import httpx
+import pytest
+import requests
+
+from jobbergate_core.auth import JobbergateAuthHandler
+from jobbergate_core.auth.exceptions import AuthenticationError
+from jobbergate_core.auth.token import TokenError, TokenType
+
+
+DUMMY_LOGIN_DOMAIN = "http://keycloak.local:8080/realms/jobbergate-local"
+DUMMY_LOGIN_AUDIENCE = "https://local.omnivector.solutions"
+DUMMY_LOGIN_CLIENT_ID = "cli"
+
+
+@pytest.fixture
+def dummy_jobbergate_auth(tmp_path):
+    """
+    Return a dummy JobbergateAuthHandler object.
+    """
+    return JobbergateAuthHandler(
+        cache_directory=tmp_path,
+        login_domain=DUMMY_LOGIN_DOMAIN,
+        login_audience=DUMMY_LOGIN_AUDIENCE,
+        login_client_id=DUMMY_LOGIN_CLIENT_ID,
+    )
+
+
+def test_auth_base_case(tmp_path, dummy_jobbergate_auth):
+    """
+    Test that the JobbergateAuthHandler class can be instantiated.
+    """
+    assert dummy_jobbergate_auth.cache_directory == tmp_path
+    assert dummy_jobbergate_auth.login_domain == DUMMY_LOGIN_DOMAIN
+    assert dummy_jobbergate_auth.login_audience == DUMMY_LOGIN_AUDIENCE
+    assert dummy_jobbergate_auth.login_client_id == DUMMY_LOGIN_CLIENT_ID
+
+    assert dummy_jobbergate_auth._access_token.content == ""
+    assert dummy_jobbergate_auth._access_token.cache_directory == tmp_path
+    assert dummy_jobbergate_auth._access_token.label == TokenType.ACCESS
+
+    assert dummy_jobbergate_auth._refresh_token.content == ""
+    assert dummy_jobbergate_auth._refresh_token.cache_directory == tmp_path
+    assert dummy_jobbergate_auth._refresh_token.label == TokenType.REFRESH
+
+
+def test_insert_token_in_request_header(respx_mock, dummy_jobbergate_auth, valid_token):
+    """
+    Test that the JobbergateAuthHandler class inserts the token in the header (performed by __call__).
+    """
+    dummy_headers = {"test-header": "test-value", "foo": "bar"}
+    expected_headers = {**dummy_headers, "Authorization": valid_token.bearer_token}
+
+    test_endpoint = "http://test.com"
+    respx_mock.get(test_endpoint).mock()
+    request = requests.Request("GET", test_endpoint, auth=dummy_jobbergate_auth, headers=dummy_headers)
+
+    assert request.headers == dummy_headers
+
+    with mock.patch.object(
+        dummy_jobbergate_auth,
+        attribute="acquire_access",
+        new=lambda *args, **kwargs: valid_token.bearer_token,
+    ):
+        prepared_request = request.prepare()
+
+    assert prepared_request.headers == expected_headers
+
+
+@pytest.mark.parametrize("procedure", (None, "load_from_cache", "refresh_tokens", "login"))
+def test_acquire_access(procedure, dummy_jobbergate_auth, valid_token):
+    """
+    Test that the acquire_tokens method works as expected.
+
+    The acquire_access method should try in this order until a valid access token is found:
+        * Get the access token stored in the instance
+        * Load the tokens from the cache directory
+        * Refresh both tokens using the refresh token.
+        * Login to fetch both tokens
+
+    This test covers all possible scenarios, making sure that the correct functions are
+    called according to the tokens available in the class.
+    """
+    expected_order_of_procedures = ("load_from_cache", "refresh_tokens", "login")
+
+    def inject_valid_access_token(*args, **kwargs):
+        dummy_jobbergate_auth._access_token = valid_token
+
+    mock_params = {key: mock.DEFAULT for key in expected_order_of_procedures}
+    if procedure is None:
+        dummy_jobbergate_auth._access_token = valid_token
+    else:
+        mock_params[procedure] = inject_valid_access_token
+
+    with mock.patch.multiple(dummy_jobbergate_auth, **mock_params) as mocked_jobbergate_auth:
+        dummy_jobbergate_auth.acquire_access()
+
+    actual_call_counter = {key: mocked.call_count for key, mocked in mocked_jobbergate_auth.items()}
+    expected_call_counter = {e: 0 for e in expected_order_of_procedures}
+    if procedure is not None:
+        actual_call_counter[procedure] = 1
+        for e in expected_order_of_procedures:
+            expected_call_counter[e] = 1
+            if e == procedure:
+                break
+
+    assert actual_call_counter == expected_call_counter
+
+
+class TestJobbergateAuthHandlerLoadFromCache:
+    """
+    Test the load_from_cache method on JobbergateAuthHandler class.
+    """
+
+    def test_no_tokens_found(self, dummy_jobbergate_auth):
+        """
+        Test that the function works as expected.
+
+        If no tokens are found in cache, the tokens dictionary should stay empty.
+        """
+        with pytest.raises(TokenError, match="Token file was not found"):
+            dummy_jobbergate_auth.load_from_cache()
+
+    def test_tokens_found__replace_loaded(self, dummy_jobbergate_auth, valid_token, expired_token):
+        """
+        Test that the function works as expected.
+
+        If a token is already loaded in the class, it should be replaced.
+        """
+        cached_tokens = {
+            TokenType.ACCESS: expired_token.replace(label=TokenType.ACCESS),
+            TokenType.REFRESH: expired_token.replace(label=TokenType.REFRESH),
+        }
+        for token in cached_tokens.values():
+            token.save_to_cache()
+
+        original_content = valid_token.content
+        expected_content = expired_token.content
+
+        dummy_jobbergate_auth.access_token = original_content
+        dummy_jobbergate_auth.refresh_token = original_content
+
+        dummy_jobbergate_auth.load_from_cache()
+
+        assert dummy_jobbergate_auth._access_token.content == expected_content
+        assert dummy_jobbergate_auth._refresh_token.content == expected_content
+
+
+def test_save_to_cache(dummy_jobbergate_auth, valid_token, expired_token):
+    """
+    Test that the save_to_cache function works as expected.
+
+    The save_to_cache function should:
+        * Create the cache directory if it does not exist
+        * Save the tokens to the cache directory
+    """
+    new_cache_directory = dummy_jobbergate_auth.cache_directory / "new_cache"
+
+    dummy_jobbergate_auth.cache_directory = new_cache_directory
+
+    access_token = dummy_jobbergate_auth._access_token.replace(
+        content=valid_token.content, cache_directory=new_cache_directory
+    )
+    dummy_jobbergate_auth._access_token = access_token
+    refresh_token = dummy_jobbergate_auth._refresh_token.replace(
+        content=expired_token.content, cache_directory=new_cache_directory
+    )
+    dummy_jobbergate_auth._refresh_token = refresh_token
+
+    assert new_cache_directory.exists() is False
+
+    dummy_jobbergate_auth.save_to_cache()
+
+    assert new_cache_directory.exists() is True
+    assert access_token.file_path.read_text() == access_token.content
+    assert refresh_token.file_path.read_text() == refresh_token.content
+
+
+class TestJobbergateAuthHandlerRefreshTokens:
+    """
+    Test the refresh_tokens method on JobbergateAuthHandler class.
+    """
+
+    def test_refresh_tokens__success(
+        self,
+        respx_mock,
+        jwt_token,
+        dummy_jobbergate_auth,
+        expired_token,
+        valid_token,
+    ):
+        """
+        Test that the function works as expected.
+        """
+        dummy_jobbergate_auth._access_token = dummy_jobbergate_auth._access_token.replace(
+            content=expired_token.content,
+        )
+        dummy_jobbergate_auth._refresh_token = dummy_jobbergate_auth._refresh_token.replace(
+            content=valid_token.content,
+        )
+
+        refreshed_access_token_content = jwt_token(custom_data="refreshed_access_token")
+        refreshed_refresh_token_content = jwt_token(custom_data="refreshed_refresh_token")
+
+        endpoint = f"{dummy_jobbergate_auth.login_domain}/protocol/openid-connect/token"
+        respx_mock.post(endpoint).mock(
+            return_value=httpx.Response(
+                httpx.codes.OK,
+                json=dict(
+                    access_token=refreshed_access_token_content,
+                    refresh_token=refreshed_refresh_token_content,
+                ),
+            ),
+        )
+
+        dummy_jobbergate_auth.refresh_tokens()
+
+        assert dummy_jobbergate_auth._access_token.content == refreshed_access_token_content
+        assert dummy_jobbergate_auth._refresh_token.content == refreshed_refresh_token_content
+
+    def test_refresh_tokens__failure_no_refresh_token(self, dummy_jobbergate_auth):
+        """
+        Test that the function raises an exception if there is no refresh token.
+        """
+        assert dummy_jobbergate_auth._refresh_token.content == ""
+        with pytest.raises(
+            AuthenticationError,
+            match="The refresh is unavailable",
+        ):
+            dummy_jobbergate_auth.refresh_tokens()
+
+    def test_refresh_tokens__failure_expired_refresh_token(self, dummy_jobbergate_auth, expired_token):
+        """
+        Test that the function raises an exception if the refresh token is expired.
+        """
+        dummy_jobbergate_auth._refresh_token = dummy_jobbergate_auth._refresh_token.replace(
+            content=expired_token.content,
+        )
+        with pytest.raises(AuthenticationError, match="Refresh token is expired"):
+            dummy_jobbergate_auth.refresh_tokens()
+
+    def test_refresh_tokens__failure_by_missing_data(
+        self,
+        respx_mock,
+        jwt_token,
+        dummy_jobbergate_auth,
+        valid_token,
+    ):
+        """
+        Test that the function raises an exception if the response is missing data.
+        """
+        dummy_jobbergate_auth._refresh_token = dummy_jobbergate_auth._refresh_token.replace(
+            content=valid_token.content,
+        )
+
+        refreshed_access_token_content = jwt_token(custom_data="refreshed_access_token")
+
+        endpoint = f"{dummy_jobbergate_auth.login_domain}/protocol/openid-connect/token"
+        respx_mock.post(endpoint).mock(
+            return_value=httpx.Response(
+                httpx.codes.OK,
+                json=dict(access_token=refreshed_access_token_content),
+            ),  # note that the refresh token is missing
+        )
+        with pytest.raises(
+            AuthenticationError,
+            match="Not all tokens were included in the response",
+        ):
+            dummy_jobbergate_auth.refresh_tokens()
+
+    def test_refresh_tokens__request_failure(
+        self,
+        respx_mock,
+        dummy_jobbergate_auth,
+        expired_token,
+        valid_token,
+    ):
+        """
+        Test that the function raises an exception if the tokens are not refreshed.
+        """
+        dummy_jobbergate_auth._refresh_token = dummy_jobbergate_auth._refresh_token.replace(
+            content=valid_token.content,
+        )
+
+        endpoint = f"{dummy_jobbergate_auth.login_domain}/protocol/openid-connect/token"
+        respx_mock.post(endpoint).mock(
+            return_value=httpx.Response(httpx.codes.BAD_REQUEST),
+        )
+
+        with pytest.raises(
+            AuthenticationError,
+            match="Unexpected error while refreshing the tokens",
+        ):
+            dummy_jobbergate_auth.refresh_tokens()
+
+
+def test_logout_success(dummy_jobbergate_auth, valid_token):
+    """
+    Test that the logout function works as expected.
+    """
+    dummy_jobbergate_auth._access_token = dummy_jobbergate_auth._access_token.replace(
+        content=valid_token.content,
+    )
+    dummy_jobbergate_auth._refresh_token = dummy_jobbergate_auth._refresh_token.replace(
+        content=valid_token.content,
+    )
+
+    access_token_path = dummy_jobbergate_auth._access_token.file_path
+    refresh_token_path = dummy_jobbergate_auth._refresh_token.file_path
+
+    dummy_jobbergate_auth.save_to_cache()
+
+    assert access_token_path.exists() is True
+    assert refresh_token_path.exists() is True
+
+    dummy_jobbergate_auth.logout()
+
+    assert dummy_jobbergate_auth._access_token.content == ""
+    assert dummy_jobbergate_auth._refresh_token.content == ""
+
+    assert access_token_path.exists() is False
+    assert refresh_token_path.exists() is False
+
+
+class TestJobbergateAuthHandlerLogin:
+    """
+    Test the login method on JobbergateAuthHandler class.
+    """
+
+    def test_login__success(self, respx_mock, dummy_jobbergate_auth, valid_token):
+        """
+        Test that the function works as expected.
+        """
+        assert dummy_jobbergate_auth._access_token.content == ""
+        assert dummy_jobbergate_auth._refresh_token.content == ""
+
+        endpoint = f"{dummy_jobbergate_auth.login_domain}/protocol/openid-connect/auth/device"
+        respx_mock.post(endpoint).mock(
+            return_value=httpx.Response(
+                httpx.codes.OK,
+                json=dict(
+                    device_code="dummy-code",
+                    verification_uri_complete="https://dummy-uri.com",
+                    interval=1,
+                    expires_in=0.5,
+                ),
+            ),
+        )
+
+        endpoint = f"{dummy_jobbergate_auth.login_domain}/protocol/openid-connect/token"
+        respx_mock.post(endpoint).mock(
+            return_value=httpx.Response(
+                httpx.codes.OK,
+                json=dict(
+                    access_token=valid_token.content,
+                    refresh_token=valid_token.content,
+                ),
+            ),
+        )
+
+        dummy_jobbergate_auth.login()
+
+        assert dummy_jobbergate_auth._access_token.content == valid_token.content
+        assert dummy_jobbergate_auth._refresh_token.content == valid_token.content
+
+    def test_login__raises_timeout(self, respx_mock, dummy_jobbergate_auth):
+        """
+        Test that the function raises an exception if the process times out.
+        """
+        endpoint = f"{dummy_jobbergate_auth.login_domain}/protocol/openid-connect/auth/device"
+        respx_mock.post(endpoint).mock(
+            return_value=httpx.Response(
+                httpx.codes.OK,
+                json=dict(
+                    device_code="dummy-code",
+                    verification_uri_complete="https://dummy-uri.com",
+                    interval=1,
+                    expires_in=1e-8,  # very small value to force timeout
+                ),
+            ),
+        )
+
+        endpoint = f"{dummy_jobbergate_auth.login_domain}/protocol/openid-connect/token"
+        respx_mock.post(endpoint).mock(
+            return_value=httpx.Response(
+                httpx.codes.BAD_REQUEST,
+            ),
+        )
+
+        with pytest.raises(AuthenticationError, match="Login expired, please try again"):
+            dummy_jobbergate_auth.login()