+++ jobbergate-test/jobbergate_test/base.py
@@ -1,0 +1,19 @@
+from abc import ABC, abstractmethod
+
+
+class BaseEntity(ABC):
+    @abstractmethod
+    def create(self):
+        pass
+
+    @abstractmethod
+    def get(self):
+        pass
+
+    @abstractmethod
+    def update(self):
+        pass
+
+    @abstractmethod
+    def list(self):
+        pass